+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 11.1.1 - Unreleased
 ### Added
-* Added ability to configure AWS Server Side Encryption algorithm for `S3S3Copier` via `copier-options.s3-sse-algorithm` configuration property.
+* Added ability to configure AWS Server Side Encryption for `S3S3Copier` via `copier-options.s3-server-side-encryption` configuration property.
 ### Fixed
 * Google FileSystem classes not being placed onto the mapreduce.application.classpath in S3MapReduceCp and DistCp mapreduce jobs
 ### Changed